% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_import.R
\name{bulk_import_delete_part}
\alias{bulk_import_delete_part}
\title{Delete bulk import part}
\usage{
bulk_import_delete_part(conn, name, part_name)
}
\arguments{
\item{conn}{\code{Td} client}

\item{name}{Bulk import session name}

\item{part_name}{Bulk import part name}
}
\value{
Return \code{TRUE} if succeeded
}
\description{
Delete bulk import part
}
\examples{
\dontrun{
conn <- Td(apikey = "xxxx")
bulk_import_delete_part(conn, sess_name, "part")
}

}
