% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_import.R
\name{bulk_import_error_records}
\alias{bulk_import_error_records}
\title{Show bulk import error records}
\usage{
bulk_import_error_records(conn, name)
}
\arguments{
\item{conn}{\code{Td} client}

\item{name}{Bulk import session name}
}
\value{
Return error records in gzipped file with msgpack stream format.
}
\description{
Show bulk import error records
}
\examples{
\dontrun{
conn <- Td(apikey = "xxxx")
bulk_import_error_records(conn, sess_name)
}

}
