% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_import.R
\name{bulk_import_upload_part}
\alias{bulk_import_upload_part}
\title{Upload bulk import part}
\usage{
bulk_import_upload_part(conn, name, part_name, file_obj)
}
\arguments{
\item{conn}{\code{Td} client}

\item{name}{Bulk import session name}

\item{part_name}{Bulk import part name}

\item{file_obj}{File connection. Should be msgpack stream with gzip compressed. Should have "time" column}
}
\value{
Return bulk import status
}
\description{
Upload bulk import part
}
\examples{
\dontrun{
conn <- Td(apikey = "xxxx")
}

}
