% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_msgpack.R
\name{check_msgpack}
\alias{check_msgpack}
\title{Check msgpack package installed or not}
\usage{
check_msgpack()
}
\description{
If msgpack package is not installed, install it from GitHub using
devtools. If it is not up to date, reinstall it.
}
