% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/td.R
\name{copy_to}
\alias{copy_to}
\title{Upload data.frame to TD}
\usage{
copy_to(dbname, table, df, overwrite = FALSE)
}
\arguments{
\item{dbname}{Target destination database name.}

\item{table}{Target table name.}

\item{df}{Input data.frame}

\item{overwrite}{Flag for overwriting the table if exists}
}
\description{
Upload data.frame to TD
}
