% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_import.R
\name{create_bulk_import}
\alias{create_bulk_import}
\title{Create bulk import}
\usage{
create_bulk_import(conn, name, dbname, table)
}
\arguments{
\item{conn}{\code{Td} client}

\item{name}{Bulk import session name}

\item{dbname}{Data base name}

\item{table}{Table name}
}
\value{
Return \code{TRUE} if succeeded
}
\description{
Create bulk import
}
\examples{
\dontrun{
conn <- Td(apikey = "xxxx")
sess_name <- uuid::UUIDgenerate()
create_bulk_import(conn, sess_name, "mydb", "mytable")
}

}
