% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{create_database}
\alias{create_database}
\title{Create a database}
\usage{
create_database(conn, dbname, params)
}
\arguments{
\item{conn}{\code{Td} client}

\item{dbname}{Target data base name}

\item{params}{Optional parameters}
}
\value{
Returns \code{TRUE} or \code{FALSE}, whether the execution succeeded or not.
}
\description{
Create a database
}
\examples{
\dontrun{
con <- Td(apikey = "xxxxx")
create_database(con, "newdb")
}

}
