% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{create_table}
\alias{create_table}
\title{Create a table}
\usage{
create_table(conn, dbname, table)
}
\arguments{
\item{conn}{\code{Td} connection}

\item{dbname}{Data base name}

\item{table}{Table name}
}
\value{
Returns \code{TRUE} or \code{FALSE}, whether the executiuon successed or not.
}
\description{
Create a table
}
\examples{
\dontrun{
conn <- Td(apikey="xxxx")
create_table(conn, "mydb", "new_table")
}

}
