% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{db_create}
\alias{db_create}
\title{Create a database}
\usage{
db_create(dbname)
}
\arguments{
\item{dbname}{Target data base name}
}
\value{
Returns \code{TRUE} or \code{FALSE}, whether the executiuon successed or not.
}
\description{
Create a database
}
\examples{
\dontrun{
db_create("newdb")
}

}
