% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{db_delete}
\alias{db_delete}
\title{Delete a database}
\usage{
db_delete(dbname)
}
\arguments{
\item{dbname}{Target data base name}
}
\value{
Returns \code{TRUE} or \code{FALSE}, whether the executiuon successed or not.
}
\description{
Delete a database
}
\examples{
\dontrun{
db_delete("mydb")
}

}
