% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{db_exists}
\alias{db_exists}
\title{Check table existence}
\usage{
db_exists(dbname)
}
\arguments{
\item{dbname}{Data base name}
}
\value{
Return \code{TRUE} or \code{FALSE}, existence
}
\description{
Check table existence
}
\examples{
\dontrun{
db_exists("mydb")
}

}
