% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_import.R
\name{delete_bulk_import}
\alias{delete_bulk_import}
\title{Delete bulk import}
\usage{
delete_bulk_import(conn, name)
}
\arguments{
\item{conn}{\code{Td} client}

\item{name}{Bulk import session name}
}
\value{
Return \code{TRUE} if succeeded
}
\description{
Delete bulk import
}
\examples{
\dontrun{
conn <- Td(apikey = "xxxx")
delete_bulk_import(conn, sess_name)
}

}
