% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{delete_database}
\alias{delete_database}
\title{Delete a database}
\usage{
delete_database(conn, dbname)
}
\arguments{
\item{conn}{\code{Td} client}

\item{dbname}{Target data base name}
}
\value{
Returns \code{TRUE} or \code{FALSE}, whether the executiuon successed or not.
}
\description{
Delete a database
}
\examples{
\dontrun{
conn <- Td(apikey="xxxx")
delete_database(conn, "mydb")
}

}
