% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{delete_table}
\alias{delete_table}
\title{Delete a table}
\usage{
delete_table(conn, dbname, table)
}
\arguments{
\item{conn}{\code{Td} connection}

\item{dbname}{Data base name}

\item{table}{Table name}
}
\value{
Returns \code{TRUE} or \code{FALSE}, whether the executiuon successed or not.
}
\description{
Delete a table
}
\examples{
\dontrun{
conn <- Td(apikey="xxxxx")
delete_table(conn, "mydb", "iris")
}

}
