% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{.as.df}
\alias{.as.df}
\title{Convert td command results into data.frame}
\usage{
.as.df(input)
}
\arguments{
\item{input}{An output of td command}
}
\description{
Convert td command results into data.frame
}
