% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{exist_database}
\alias{exist_database}
\title{Check table existence}
\usage{
exist_database(conn, dbname)
}
\arguments{
\item{conn}{\code{Td} client}

\item{dbname}{Data base name}
}
\value{
Return \code{TRUE} or \code{FALSE}, existence
}
\description{
Check table existence
}
\examples{
\dontrun{
conn <- Td(apikey="xxxx")
exist_database(conn, "mydb")
}

}
