% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_import.R
\name{freeze_bulk_import}
\alias{freeze_bulk_import}
\title{Freeze bulk import part}
\usage{
freeze_bulk_import(conn, name)
}
\arguments{
\item{conn}{\code{Td} client}

\item{name}{Bulk import session name}
}
\value{
Return \code{TRUE} if succeeded
}
\description{
Freeze bulk import part
}
\examples{
\dontrun{
conn <- Td(apikey = "xxxx")
freeze_bulk_import(conn, sess_name)
}

}
