% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_auto}
\alias{import_auto}
\title{Import file to TD. Equivalent with td import:auto}
\usage{
import_auto(dest, file_name, delimiter, column_types, time_value,
  format = "csv", header = TRUE)
}
\arguments{
\item{dest}{Destination for TD. `db_name.table_name`}

\item{file_name}{Input file name to upload}

\item{delimiter}{CSV/TSV option. delimiter CHAR; default="," at csv, TAB at tsv}

\item{column_types}{column types [string, int, long, double]}

\item{time_value}{time column's value.}

\item{format}{source file format [csv, tsv, json, msgpack, apache, regex, mysql]; default=csv}

\item{header}{CSV/TSV option. first line includes column names}

\item{quote}{CSV/TSV option. [DOUBLE, SINGLE, NONE]; if csv format, default=DOUBLE. if tsv format, default=NONE}
}
\description{
Import file to TD. Equivalent with td import:auto
}
