% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_msgpack.R
\name{install_msgpack}
\alias{install_msgpack}
\title{Install the msgpack package after checking with the user}
\usage{
install_msgpack()
}
\description{
Install the msgpack package after checking with the user
}
