% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_import.R
\name{list_bulk_import_parts}
\alias{list_bulk_import_parts}
\title{List bulk import parts}
\usage{
list_bulk_import_parts(conn, name)
}
\arguments{
\item{conn}{\code{Td} client}

\item{name}{Bulk import session name}
}
\value{
Return bulk import parts list
}
\description{
List bulk import parts
}
\examples{
\dontrun{
conn <- Td(apikey = "xxxx")
list_bulk_import_parts(conn, sess_name)
}

}
