% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{list_databases}
\alias{list_databases}
\title{Show database list}
\usage{
list_databases(conn)
}
\arguments{
\item{conn}{\code{Td} connection}
}
\value{
Returns a \code{data.frame} of the database list
}
\description{
Show database list
}
\examples{
\dontrun{
conn <- Td(apikey="xxxx")
list_databases(conn)
}

}
