% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{list_tables}
\alias{list_tables}
\title{Show list of tables}
\usage{
list_tables(conn, dbname)
}
\arguments{
\item{conn}{\code{Td} connection}

\item{dbname}{Data base name. Optional, but highly recommended to prevent timeout.}
}
\value{
Returns a \code{data.frame} of a list of tables or \code{FALSE} if not exists.
}
\description{
Show list of tables
}
\examples{
\dontrun{
conn <- Td(apikey="xxxxx")
list_tables(conn, "mydb")
}

}
