% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_import.R
\name{show_bulk_import}
\alias{show_bulk_import}
\title{Show bulk import}
\usage{
show_bulk_import(conn, name)
}
\arguments{
\item{conn}{\code{Td} client}

\item{name}{Bulk import session name}
}
\value{
Return bulk import status
}
\description{
Show bulk import
}
\examples{
\dontrun{
conn <- Td(apikey = "xxxx")
show_bulk_import(conn, sess_name)
}

}
