% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{table_create}
\alias{table_create}
\title{Create a table}
\usage{
table_create(dbname, table)
}
\arguments{
\item{dbname}{Data base name}

\item{table}{Table name}
}
\value{
Returns \code{TRUE} or \code{FALSE}, whether the executiuon successed or not.
}
\description{
Create a table
}
\examples{
\dontrun{
table_create("mydb", "new_table")
}

}
