% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{table_delete}
\alias{table_delete}
\title{Delete a table}
\usage{
table_delete(dbname, table)
}
\arguments{
\item{dbname}{Data base name}

\item{table}{Table name}
}
\value{
Returns \code{TRUE} or \code{FALSE}, whether the executiuon successed or not.
}
\description{
Delete a table
}
\examples{
\dontrun{
table_delete("mydb", "iris")
}

}
