% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{table_exists}
\alias{table_exists}
\title{Check table existence}
\usage{
table_exists(dbname, table)
}
\arguments{
\item{dbname}{Data base name}

\item{table}{Table name}
}
\description{
Check table existence
}
