% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{td_execute}
\alias{td_execute}
\title{Execute td command}
\usage{
td_execute(command, opts, format = FALSE, intern = TRUE,
  quiet = FALSE, timeout = FALSE)
}
\arguments{
\item{command}{First argument for td command. e.g. db:list}

\item{opts}{Options for td command. \code{carachter} or \code{vector} of \code{character}.}

\item{format}{Flag for parsing td command output as TSV.}

\item{intern}{Flag for intern the output of system function.}

\item{quiet}{Suppress warning for td command.}

\item{timeout}{Timeout seconds for executing td command.}
}
\description{
Execute td command
}
\examples{
\dontrun{
# With character for \\code{opts}
td_execute("table:list", "mydb", format = TRUE)

# With character vector for \\code{opts}
td_execute("table:create", c("mydb", "iris"))
}

}
