% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{update_schema}
\alias{update_schema}
\title{Update schema of a table}
\usage{
update_schema(conn, dbname, table, schema)
}
\arguments{
\item{conn}{\code{Td} connection}

\item{dbname}{Data base name}

\item{table}{Table name}

\item{schema}{Schema of the table to be updated}
}
\value{
Returns \code{TRUE} or \code{FALSE}, whether the execution succeeded or not.
}
\description{
Update schema of a table
}
\examples{
\dontrun{
conn <- Td(apikey = "xxxxx")
s <- rbind(
c("sepal_length", "double", "sepal_length"),
c("sepal_width", "double", "sepal_width"),
c("petal_length", "double", "petal_length"),
c("petal_width", "double", "petal_width"),
c("species", "string", "species"))
udpate_schema(conn, "mydb", "iris", s)
}

}
