% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_import.R
\name{wait_bulk_import}
\alias{wait_bulk_import}
\title{Wait bulk import until finished}
\usage{
wait_bulk_import(conn, sess_name)
}
\arguments{
\item{conn}{\code{Td} client}

\item{sess_name}{Bulk import session name}
}
\description{
Wait bulk import until finished
}
\examples{
\dontrun{
conn <- Td(apikey = "xxxx")
wait_bulk_import(conn, sess_name)
}

}
