% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{db_show}
\alias{db_show}
\title{Describe information of a database}
\usage{
db_show(dbname, quiet = FALSE)
}
\arguments{
\item{dbname}{Target data base name}

\item{quiet}{Suppress td command warning}
}
\description{
Describe information of a database
}
