% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{table_show}
\alias{table_show}
\title{Describe information of a table}
\usage{
table_show(dbname, table, quiet = FALSE)
}
\arguments{
\item{dbname}{Data base name}

\item{table}{Table name}

\item{quiet}{Suppress warning for td command}
}
\value{
A \code{list} of table info or \code{FALSE} if not exists.
}
\description{
Describe information of a table
}
\examples{
\dontrun{
table_show("mydb", "iris")
}

}
