% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TdClient.R
\name{Td}
\alias{Td}
\title{Td client}
\usage{
Td(endpoint, apikey, user_agent, headers)
}
\arguments{
\item{endpoint}{Endpoint to TD API}

\item{apikey}{API key for TD}

\item{user_agent}{User-Agent. optional}

\item{headers}{Default headres. optional}
}
\description{
Td client
}
