% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TdClient.R
\name{Td}
\alias{Td}
\title{Connect to TD}
\usage{
Td(endpoint, apikey, user_agent, headers, http_proxy = NULL)
}
\arguments{
\item{endpoint}{Endpoint to TD API}

\item{apikey}{API key for TD}

\item{user_agent}{User-Agent. optional}

\item{headers}{Default headres. optional}

\item{http_proxy}{HTTP proxy setting. optional}
}
\description{
Connect to TD
}
