% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/td.R
\name{td_upload}
\alias{td_upload}
\title{Upload data.frame to TD}
\usage{
td_upload(dbname, table, df, overwrite = FALSE)
}
\arguments{
\item{dbname}{Target destination database name.}

\item{table}{Target table name.}

\item{df}{Input data.frame.}

\item{overwrite}{Flag for overwriting the table if exists. It doesn't overwrite database.}
}
\description{
Upload data.frame to TD
}
